+++ nodestream/project/pipeline_scope.py
@@ -1,9 +1,8 @@
 from importlib import resources
 from typing import Dict, Iterable, List
 
-from nodestream.pipeline.scope_config import ScopeConfig
-
 from ..file_io import LoadsFromYaml, SavesToYaml
+from ..pipeline.scope_config import ScopeConfig
 from ..schema.schema import (
     AggregatedIntrospectiveIngestionComponent,
     IntrospectiveIngestionComponent,
